+++ apps/api/src/common/index.ts
@@ -13,6 +13,8 @@ export { AllExceptionsFilter } from "./filters/http-exception.filter";
 export { PaginationQueryDto } from "./dto/pagination.dto";
 export { paginationArgs, paginatedResponse } from "./helpers/paginate";
 export { sanitizeFilename, contentDisposition } from "./utils/sanitize";
+export { toCsv } from "./utils/csv";
+export type { CsvColumn } from "./utils/csv";
 export { assertProjectAccess } from "./helpers/assert-project-access";
 export { BLOCKED_EXTENSIONS } from "./utils/blocked-extensions";
 export type {